OverHaul of Timing Extraction
Timing Extraction of h5 files now creates an array of the start of each frame.  for each trial it looks for frames that were between that trial start and end. This should work for all trial types

now checks there are enough trial onsets and offsets found.

the first frame is now the closest frame to the onset of the trial, decreasing potential trial jitter from 1 frame to a 1/2 frame.

if run in debug mode, added a plot of trial with frame timing, trial timing and how the first trial changes between old and new extractions if it differred.